Cometics - Test's display name in english - a better way to use predefined data ( @Sql )
