+++ src/test/resources/test-data.sql
@@ -1,0 +1,4 @@
+DELETE FROM task;
+INSERT INTO task (id, title, description, status, user_id) VALUES (999, 'Task 1', 'Task 1', 'IN_PROGRESS', 'task1');
+INSERT INTO task (id, title, description, status, user_id) VALUES (998, 'Task 2', 'Task 2', 'COMPLETED', 'task2');
+INSERT INTO task (id, title, description, status, user_id) VALUES (997, 'Task 3', 'Task 3', 'CANCELLED', 'task3');
